Move env registration to mujoco.py
setup.py currently imports src.benchmark_environments to get the
version, so src.benchmark_environments should not import gym, since it
might be unavailable when setup.py starts (in which case installation
then fails). Thus we move for now the environment registration to
mujoco.py.
